# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZeroHunger Enterprise SaaS API",
    "version": "2.0.0",
    "description": "\n## ZeroHunger Enterprise Food Rescue & Redistribution Network API\n\nZeroHunger is an AI-powered logistics platform connecting commercial food donors (hotels, restaurants, bakeries) with verified NGOs, shelters, and volunteer couriers to eliminate food waste and hunger.\n\n### Key Capabilities\n* **Authentication & Identity**: JWT bearer auth, role-based access control, OTP verification, self-service profile setup.\n* **Food Rescue Listings**: Post surplus food, category classification, expiration window tracking, geolocation pickup bounds.\n* **Fulfillment & Dispatch**: Bulk request placement by NGOs, volunteer task allocation, delivery status pipeline.\n* **Enterprise Analytics**: Real-time stats, weekly trends, environmental CO₂ impact metrics, and administrative audit logs.\n\n### Security\nAll protected endpoints require an HTTP Authorization header formatted as:\n```\nAuthorization: Bearer <your_jwt_token>\n```\n    ",
    "termsOfService": "https://zerohunger.org/terms",
    "contact": {
      "name": "ZeroHunger API Support Team",
      "email": "api-support@zerohunger.org",
      "url": "https://zerohunger.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide a valid JWT token obtained from `/api/v1/auth/login` or `/api/v1/auth/register`."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "66a9b12c8f413d29a01048b1"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@grandhyatt.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "restaurant",
              "donor",
              "ngo",
              "volunteer",
              "admin"
            ],
            "example": "restaurant"
          },
          "city": {
            "type": "string",
            "example": "San Francisco"
          },
          "address": {
            "type": "string",
            "example": "123 Market St, Dock 4"
          },
          "phone": {
            "type": "string",
            "example": "+14155552671"
          },
          "organizationName": {
            "type": "string",
            "example": "Grand Hyatt Kitchens"
          },
          "bio": {
            "type": "string",
            "example": "Executive culinary team dedicated to zero food waste."
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-01T10:00:00.000Z"
          }
        }
      },
      "Food": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "66a9c80f8f413d29a01049c5"
          },
          "title": {
            "type": "string",
            "example": "50 Fresh Gourmet Prepared Meals"
          },
          "category": {
            "type": "string",
            "enum": [
              "cooked",
              "raw",
              "packaged",
              "beverage",
              "bakery",
              "dairy",
              "other"
            ],
            "example": "cooked"
          },
          "quantity": {
            "type": "string",
            "example": "50 boxes"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "reserved",
              "collected",
              "expired"
            ],
            "example": "available"
          },
          "expiryTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-03T18:00:00.000Z"
          },
          "pickupTime": {
            "type": "string",
            "example": "Today between 5 PM - 8 PM"
          },
          "city": {
            "type": "string",
            "example": "San Francisco"
          },
          "pickupAddress": {
            "type": "string",
            "example": "123 Market St, Dock 4"
          },
          "description": {
            "type": "string",
            "example": "Surplus dinner meals packaged in thermal containers."
          },
          "donatedBy": {
            "$ref": "#/components/schemas/User"
          },
          "claimedBy": {
            "$ref": "#/components/schemas/User"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "/uploads/food-102.jpg"
            ]
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-02T01:00:00.000Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Resource not found or unauthorized"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Invalid resource identifier"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Must be a valid email address"
                }
              }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 45
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication token is missing or invalid (401)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required. Token missing or expired."
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient role permissions (403)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Access forbidden. Required role permission missing."
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Requested endpoint or resource not found (404)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Resource not found"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request payload validation errors (400 / 422)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/food": {
      "get": {
        "summary": "Discover surplus food rescue listings",
        "description": "Retrieve active surplus food listings filtered by category, status, city, or search term.",
        "tags": [
          "Food Donations"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search keyword for title, city, or donor name"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "cooked",
                "raw",
                "packaged",
                "beverage",
                "bakery",
                "dairy",
                "other"
              ]
            },
            "description": "Filter by food category"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "available",
                "reserved",
                "collected",
                "expired"
              ]
            },
            "description": "Filter by availability status"
          },
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string"
            },
            "description": "Filter listings by city"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of food listings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Food"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post a new surplus food donation",
        "description": "Publish a new surplus food listing. Requires authenticated donor or restaurant role. Supports optional multipart image attachments.",
        "tags": [
          "Food Donations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "quantity",
                  "expiryTime"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "50 Fresh Gourmet Dinner Boxes"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "cooked",
                      "raw",
                      "packaged",
                      "beverage",
                      "bakery",
                      "dairy",
                      "other"
                    ],
                    "example": "cooked"
                  },
                  "quantity": {
                    "type": "string",
                    "example": "50 boxes"
                  },
                  "expiryTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-08-05T18:00:00.000Z"
                  },
                  "pickupTime": {
                    "type": "string",
                    "example": "Today between 5 PM - 8 PM"
                  },
                  "city": {
                    "type": "string",
                    "example": "San Francisco"
                  },
                  "pickupAddress": {
                    "type": "string",
                    "example": "123 Market St, Dock 4"
                  },
                  "description": {
                    "type": "string",
                    "example": "Surplus dinner meals stored in thermal food containers."
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "quantity",
                  "expiryTime"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "string"
                  },
                  "expiryTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "pickupTime": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "pickupAddress": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Food donation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/food/me": {
      "get": {
        "summary": "Retrieve user's posted food listings",
        "description": "Returns all surplus food donations posted by the authenticated user.",
        "tags": [
          "Food Donations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User's food donations retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Food"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/v1/food/{id}": {
      "get": {
        "summary": "Get single food listing by ID",
        "description": "Retrieve detailed metadata for a specific food rescue listing.",
        "tags": [
          "Food Donations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB Food Document ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Food listing detail retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "summary": "Update food listing details",
        "description": "Update title, category, quantity, expiry time, or status of an existing food listing.",
        "tags": [
          "Food Donations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "available",
                      "reserved",
                      "collected",
                      "expired"
                    ]
                  },
                  "quantity": {
                    "type": "string"
                  },
                  "expiryTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Food listing updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "summary": "Delete a food listing",
        "description": "Permanently remove a food donation listing.",
        "tags": [
          "Food Donations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Food listing deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  }
}